refactor(commands): centralize autocomplete logic in utility function
The autocomplete logic for command interactions has been moved to a
utility function 'autoComplete'. This reduces code repetition across
the 'display' and 'edit' command modules, improving maintainability
and readability.
